Refactor raga components
