+++ pages/index.tsx
@@ -1,5 +1,5 @@
 import { Welcome } from '../library/components/Welcome/Welcome';
-import RagaExplore from '../library/components/RagaExplore/RagaExplore'
+import RagaExplore from '../library/components/raga_components/RagaExplore/RagaExplore'
 
 export default function HomePage() {
   return (
